% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convFunctionCalls.R
\name{makeFuncMaps}
\alias{makeFuncMaps}
\title{Turn dictionary lines into functions that map matlab to R function calls}
\usage{
makeFuncMaps(addDict = NULL, pathDict = "")
}
\arguments{
\item{addDict}{An optional character vector with manufactored lines}

\item{pathDict}{The path to a text file with the dictionry lines written to it}
}
\value{
a list of functions with names of matlab functions and the function
to change the matlab input into an R function call
}
\description{
Turn dictionary lines into functions that map matlab to R function calls
}

