% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mat2r.R
\name{mat2r}
\alias{mat2r}
\title{mat2r}
\usage{
mat2r(pathInMat, pathOutR, funcConverters = NULL, verbose = 1)
}
\arguments{
\item{pathInMat}{A string file path with the input MatLab / Octave code to be converted}

\item{pathOutR}{A string file path with the desired output file location}

\item{funcConverters}{a list of function converters that the user wants ot use this conversion}

\item{verbose}{A number indicating the amount of output desired. (0-2)}
}
\value{
A list containing the report file and the converted code.
}
\description{
The top level driver function to call the converting functions and handle the
input and output.
}

