% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convData.R
\name{makeDataMap}
\alias{makeDataMap}
\title{Make the maps for the data}
\usage{
makeDataMap(leftSym, rightSym, rClass)
}
\arguments{
\item{leftSym}{The left symbol that contains the matlab data}

\item{rightSym}{the right symbol that contains the matlab data}

\item{rClass}{The formal r class name that defines what the R data is
  outputted as}
}
\value{
A function that takes in a matlab lines and changes the data into R
  data lines
}
\description{
Make the maps for the data
}

